Downloading data from Ibama database

This function can be used to get datasets about fines applied or collected by the Brazilian Institute for the Environment and Renewable Natural Resources (IBAMA). This function allows to search datasets for different states of Brasil, or for the whole country. The resulting dataset is cleaned.

get_dataset_ibamam(dataset, uf = "all", clean = TRUE)

Arguments

dataset
Which dataset to search for. Can receive one of the values: for fines applied, use 'distribuidas', or for fines collected, use 'arrecadadas'.

uf
Abreviation of the name of the states (two letters in lower case) that data will be obtained. Default is "all", so all data from all the states of Brazil will be returned by default.

clean
Whether the dataset should be cleaned or not. If TRUE, the dataset will be cleaned. If FALSE, the returned dataset will be the original version, without modifications. Default is TRUE.

Value

a data frame:

dataAuto

Date of issue of the tax assessment notice

nomeMunicipio

Name of the municipality cleaned from the original dataset. Only for clean = TRUE

municipio

Name of the municipality in the original dataset. Only for clean = FALSE

nomeMunicipio_geobr

Name of the municipality used to join with the dataset of geobr. Only for clean = TRUE

codigoMunicipio

IBGE code of the municipality in the dataset of geobr. Only for clean = TRUE

dataPagamento

Date of payment of the infraction. Only for dataset = "arrecadadas"

valorPago

Amount of money paid up to the date of the last update of the report. Only for dataset = "arrecadadas"

numAI

Identification number of the infraction notice

tipoInfracao

Type of infraction committed

ultimaAtualizacaoRelatorio

Date when the data was last updated

parcela

Installment number. Only for dataset = "arrecadadas"

quantidadeParcela

Total number of payments in installments. Only for dataset = "arrecadadas"

uf

State of the federation

tipoAuto

Category of the occurrence (Ex: fine, daily fine, warning)

moeda

Money currency used in Brazil when the

statusDebito

Monetary status of the infraction

enquadramentoLegal

Which legal device framing the infraction notice

nomeRazaoSocial

Name of the person or company that received the notice

cpfCnpj

Number of CPF or CNPJ of the person or company that received the notice

valorbaseParcela

Value of the installments for payment of the auto value. Only for dataset = "arrecadadas"

valorAuto

Monetary amount of the tax assessment notice

enquadramentoJuridico

Which type of document the cpfCnpj refears to: CPF or CNPJ. Only for clean = TRUE

Examples

if (FALSE) {
get_dataset_ibamam("distribuidas", "SE")
get_dataset_ibamam("arrecadadas", "SE")
}