Downloading data from Ibama database

This function can be used to get datasets about fines applied or collected by the Brazilian Institute for the Environment and Renewable Natural Resources (IBAMA). This function allows to search datasets for different states of Brasil, or for the whole country. The resulting dataset is cleaned.

get_dataset_ibamam(dataset, uf = "all", clean = TRUE)

Arguments

dataset
Which dataset to search for. Can receive one of the values: for fines applied, use 'distribuidas', or for fines collected, use 'arrecadadas'.

uf
Abreviation of the name of the states (two letters in lower case) that data will be obtained. Default is "all", so all data from all the states of Brazil will be returned by default.

clean
Whether the dataset should be cleaned or not. If TRUE, the dataset will be cleaned. If FALSE, the returned dataset will be the original version, without modifications. Default is TRUE.

Value

a data frame:

dataAuto

Date of issue of the tax assessment notice

nomeMunicipio

Name of the municipality cleaned from the original dataset. Only for clean = TRUE

municipio

Name of the municipality in the original dataset. Only for clean = FALSE

nomeMunicipio_geobr

Name of the municipality used to join with the dataset of geobr. Only for clean = TRUE

codigoMunicipio

IBGE code of the municipality in the dataset of geobr. Only for clean = TRUE

dataPagamento

Date of payment of the infraction. Only for dataset = "arrecadadas"

valorPago

Amount of money paid up to the date of the last update of the report. Only for dataset = "arrecadadas"

numAI

Identification number of the infraction notice

tipoInfracao

Type of infraction committed

ultimaAtualizacaoRelatorio

Date when the data was last updated

parcela

Installment number. Only for dataset = "arrecadadas"

quantidadeParcela

Total number of payments in installments. Only for dataset = "arrecadadas"

uf

State of the federation

tipoAuto

Category of the occurrence (Ex: fine, daily fine, warning)

moeda

Money currency used in Brazil when the

statusDebito

Monetary status of the infraction

enquadramentoLegal

Which legal device framing the infraction notice

nomeRazaoSocial

Name of the person or company that received the notice

cpfCnpj

Number of CPF or CNPJ of the person or company that received the notice

valorbaseParcela

Value of the installments for payment of the auto value. Only for dataset = "arrecadadas"

valorAuto

Monetary amount of the tax assessment notice

enquadramentoJuridico

Which type of document the cpfCnpj refears to: CPF or CNPJ. Only for clean = TRUE

Examples

get_dataset_ibamam("distribuidas", "SE")

#> # A tibble: 3,886 x 16
#>    dataAuto   nomeMunicipio  nomeMunicipio_g… codigoMunicipio numAI tipoInfracao
#>    <date>     <chr>          <chr>            <chr>           <chr> <fct>       
#>  1 2020-10-14 INDIAROBA      Indiaroba        2802809         "AAD… Flora       
#>  2 2020-05-28 LAGARTO        Lagarto          2803500         "A2O… Flora       
#>  3 2021-04-20 AREIA BRANCA   Areia Branca     2800506         "A4K… Fauna       
#>  4 2021-02-27 AREIA BRANCA   Areia Branca     2800506         "A9V… Flora       
#>  5 2020-09-09 ARACAJU        Aracaju          2800308         "CI5… Fauna       
#>  6 2019-11-01 LAGARTO        Lagarto          2803500         "CUC… Flora       
#>  7 2021-03-17 BREJO GRANDE   Brejo Grande     2800704         "C3B… Outras      
#>  8 2020-07-26 ARACAJU        Aracaju          2800308         "DKD… Fauna       
#>  9 2020-07-26 ARACAJU        Aracaju          2800308         "DVB… Fauna       
#> 10 2020-04-02 PORTO DA FOLHA Porto da Folha   2805604         "D0Q… Flora       
#> # … with 3,876 more rows, and 10 more variables:
#> #   ultimaAtualizacaoRelatorio <dttm>, uf <chr>, situacaoDebito <chr>,
#> #   tipoAuto <fct>, moeda <fct>, enquadramentoLegal <fct>,
#> #   nomeRazaoSocial <chr>, cpfCnpj <chr>, valorAuto <dbl>,
#> #   enquadramentoJuridico <fct>

get_dataset_ibamam("arrecadadas", "SE")

#> # A tibble: 9,529 x 21
#>    dataAuto   nomeMunicipio     nomeMunicipio_geo… codigoMunicipio dataPagamento
#>    <date>     <chr>             <chr>              <chr>           <date>       
#>  1 2014-06-03 NOSSA SENHORA DO… Nossa Senhora do … 2804805         2014-07-07   
#>  2 1996-02-27 ESTANCIA          Estância           2802106         1996-03-08   
#>  3 1996-11-19 UMBAUBA           Umbaúba            2807600         1997-09-26   
#>  4 1996-11-19 UMBAUBA           Umbaúba            2807600         1997-01-13   
#>  5 1996-11-26 ESTANCIA          Estância           2802106         1998-08-10   
#>  6 1996-11-26 ESTANCIA          Estância           2802106         2009-07-06   
#>  7 1997-03-06 ESTANCIA          Estância           2802106         1998-07-10   
#>  8 1997-03-06 ESTANCIA          Estância           2802106         1998-08-10   
#>  9 1997-03-06 ESTANCIA          Estância           2802106         1998-10-09   
#> 10 1997-03-06 ESTANCIA          Estância           2802106         1998-02-10   
#> # … with 9,519 more rows, and 16 more variables: valorPago <dbl>, numAI <chr>,
#> #   tipoInfracao <fct>, ultimaAtualizacaoRelatorio <dttm>, parcela <int>,
#> #   quantidadeParcela <int>, uf <chr>, tipoAuto <fct>, moeda <fct>,
#> #   statusDebito <chr>, enquadramentoLegal <fct>, nomeRazaoSocial <chr>,
#> #   cpfCnpj <chr>, valorbaseParcela <dbl>, valorAuto <dbl>,
#> #   enquadramentoJuridico <fct>